network: Add automatic source routing support for DHCP/ipv6RA addrs
When there are multiple network interfaces on a machine attached to the same
network, Linux needs some help to work out when traffic should go out of each
interface, since addresses on Linux belong to the host, not the interface.

One way to configure this is with "source routing". In this scheme, you have
separate routing tables for each interface, each of which has an entry to send
traffic for the same local network out via that interface. Then, you use
routing policy rules to configure which routing table to use based on the
source address of a packet.

For example, if you had two interfaces eth1 and eth2 connected to the same
network, with addresses 10.0.0.2/24 and 10.0.0.3/24, and you wanted traffic
to use eth1 _unless_ its part of a TCP flow specifically directed at 10.0.0.3,
you might have route tables like so:

$ ip route list table 100
default via 10.0.0.1 dev eth1 proto dhcp src 10.0.0.2  metric 1024
10.0.0.0/24 dev eth1 proto dhcp scope link src 10.0.0.2
10.0.0.2 dev eth1 proto dhcp scope link src 10.0.0.2 metric 1024

$ ip route list table 101
default via 10.0.0.1 dev eth2 proto dhcp src 10.0.0.3  metric 1024
10.0.0.0/24 dev eth2 proto dhcp scope link src 10.0.0.3
10.0.0.3 dev eth2 proto dhcp scope link src 10.0.0.3 metric 1024

$ ip rule list
0:	from all lookup local
100:	from 10.0.0.2/24 lookup 100
101:	from 10.0.0./24 lookup 101
200:    from all lookup 100
32766:	from all lookup main
32767:	from all lookup default

Systemd already has support both for configuring DHCP routes in a
user-specified route table, and for configuring routing policy rules to achieve
this kind of source routing. However, there is no easy way to set this up in a
way that the addresses in the routing policy rules are automatically obtained
from DHCP/IPv6 router advertisements. That is what this patch tries to solve.

To use it in the scenario above, you would configure each interface like so:

[Match]
Name=eth1

[Network]
DHCP=yes

[DHCPv4]
RouteTable=100
EnableSourceRoutingRules=yes
SourceRoutingRulesPriority=100

[RoutingPolicyRule]
Priority=200
Table=100
Family=ipv4

The EnableSourceRoutingRules directive cause routing rules to be created for
the DHCP address on the network interface. For eth1 in our example above, we would
also want to create two RoutingPolicyRule entries to creat the "from all lookup
100" entry as well, so that outgoing traffic uses this interface. We would
_not_ want this for eth2.